Change package and Maven groupId to app.cash
CashApp Inflation Inject by CashApp™
