+++ gradle.properties
@@ -1,19 +1,19 @@
-GROUP=com.squareup.inject
+GROUP=app.cash.inject
 VERSION_NAME=1.0.0-SNAPSHOT
 
 POM_DESCRIPTION=Constructor-inject views during layout inflation
 
-POM_URL=https://github.com/square/InflationInjection/
-POM_SCM_URL=https://github.com/square/InflationInjection/
-POM_SCM_CONNECTION=scm:git:git://github.com/square/InflationInjection.git
-POM_SCM_DEV_CONNECTION=scm:git:ssh://git@github.com/square/InflationInjection.git
+POM_URL=https://github.com/cashapp/InflationInjection/
+POM_SCM_URL=https://github.com/cashapp/InflationInjection/
+POM_SCM_CONNECTION=scm:git:git://github.com/cashapp/InflationInjection.git
+POM_SCM_DEV_CONNECTION=scm:git:ssh://git@github.com/cashapp/InflationInjection.git
 
 POM_LICENCE_NAME=The Apache Software License, Version 2.0
 POM_LICENCE_URL=http://www.apache.org/licenses/LICENSE-2.0.txt
 POM_LICENCE_DIST=repo
 
-POM_DEVELOPER_ID=square
-POM_DEVELOPER_NAME=Square, Inc.
+POM_DEVELOPER_ID=cashapp
+POM_DEVELOPER_NAME=CashApp
 
 # Disable R8's ProGuard compatibility mode.
 android.enableR8.fullMode=true
