+++ inflation-inject/src/main/java/app/cash/inject/inflation/Inflated.java
@@ -1,0 +1,12 @@
+package app.cash.inject.inflation;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import static java.lang.annotation.ElementType.PARAMETER;
+import static java.lang.annotation.RetentionPolicy.CLASS;
+
+@Retention(CLASS)
+@Target(PARAMETER)
+public @interface Inflated {
+}
