+++ inflation-inject/src/main/java/app/cash/inject/inflation/InflationModule.java
@@ -1,0 +1,12 @@
+package app.cash.inject.inflation;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.RetentionPolicy.CLASS;
+
+@Target(TYPE)
+@Retention(CLASS)
+public @interface InflationModule {
+}
