+++ inflation-inject/src/main/java/app/cash/inject/inflation/ViewFactory.java
@@ -1,0 +1,12 @@
+package app.cash.inject.inflation;
+
+import android.content.Context;
+import android.util.AttributeSet;
+import android.view.View;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
+public interface ViewFactory {
+  @NonNull
+  View create(@NonNull Context context, @Nullable AttributeSet attrs);
+}
