--- inflation-inject/src/main/java/com/squareup/inject/inflation/InflationInject.java
@@ -1,12 +1,0 @@
-package com.squareup.inject.inflation;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.CONSTRUCTOR;
-import static java.lang.annotation.RetentionPolicy.CLASS;
-
-@Retention(CLASS)
-@Target(CONSTRUCTOR)
-public @interface InflationInject {
-}
