--- inflation-inject/src/main/java/com/squareup/inject/inflation/InflationModule.java
@@ -1,12 +1,0 @@
-package com.squareup.inject.inflation;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.TYPE;
-import static java.lang.annotation.RetentionPolicy.CLASS;
-
-@Target(TYPE)
-@Retention(CLASS)
-public @interface InflationModule {
-}
